Trade rdbms for custom resources (almost)
We've decided that a custom resource data model will let k8s do more
of the heavy lifting than an rdbms model. This commit rips out the
database model, and all the web service stuff that the control plane
doesn't need, and adds controllers for the egw-resource-model custom
resources.

It's not quite there yet, though: we've got the controllers and we've
got the xDS control plane but they're not yet linked together.