Use a namespace per account
K8s enforces uniqueness on "namespace/objectName" so if we want two
customers to be able to use the same ServiceGroup name we need to use
namespaces.

The scheme is described in the README but tl;dr: there's an "egw"
namespace that acts as an index. Accounts are stored there, so we can
look them up in one place.  Each account has its own namespace called
"egw-accountName".  Service groups and load balancers are stored
there.

This should give us the flexibility that we need without too many
unneeded complications.