Switch from static Envoy node-id to dynamic based on lb name
To serve configuration to multiple Envoy instances we need a scheme to
have each one get the proper config. We were hard-coding a single
node-id but now the configuration node-id is based on the loadbalancer
custom resource's:

service.ObjectMeta.Namespace + "/" + service.ObjectMeta.Name

The EGW operator will need to pass the appropriate node-id to each
Envoy when it launches its pod.