Fill in some previously hard-coded stuff
We can now look up most of what used to be hard-coded, the service
prefix for example.

The name of the multus interface is still hard-coded to "multus0" but
at least the code that creates the interface and the code that uses it
are in the same module so they share a common definition.