Fix iptables setup
The iptables command was crashing because it couldn't write its lock
file to /run/xtables.lock because the root filesystem was read-only. I
mounted an emptyDir filesystem to /run to fix that.

Then I was getting an error from iptables not being able to parse the
command line. The command-line definition combined some parameters
into single arguments which likely works with some shells and not
others. I changed the args so there's now one CLI argument per
argument to exec.Command().

I also simplified the error handling so the error object returned
from the subcommands gets returned to the caller.