Get multus bridge interface name from configuration
It used to be hard-coded to "multus0" and that's OK most of the time
but sometimes we'll want to override that so we now get the name from
the service prefix.

As part of this I moved the bridge setup code from main.go to the
service prefix controller since we can't set up the interface until we
know what its name is. This should also allow us to have multiple
multus bridge interfaces if we need to.