Configure bridge even if netattachdef exists
We were using the existence of a NetworkAttachmentDefinition as a
flag, i.e., if it existed we assumed that everything was good and we
didn't need to configured anything. That's not quite right. The
netattachdef is more durable that the underlying linux network config
because the netattachdef will survive a node reboot while the
underlying linux config won't.

Now we check the existence of a netattachdef and if we don't find one
we create it. In either case, though, we configure the linux network,
too.