Initialize the multus bridge intf in the loadbalancer controller
This fixes a race condition bug that happens when the cluster is
coming up from a reboot. In that case the netattachdef is
present (because k8s/etcd is durable) but the linux networking config
is wiped clean so there's no multus0 bridge.

If the serviceprefix event happens first then everything's ok but
there's no guarantee and if the loadbalancer events happens first it
used to fail because it assumed that the multus bridge was ready to be
configured.

Now we create/configure the multus bridge in the loadbalancer
controller so we're sure that it's there when we need it.