Use envsubst to set the deployment docker image
We were using the scaffolded approach that uses kustomize to set the
image and has the Makefile explicitly modify the image in the
kustomization.yaml file every time it runs. That doesn't work well for
us, though, because it makes it very easy to commit a change that
switches the image to your "-dev" image for everyone.

We want to set the image dynamically so we're not modifying
git-controlled files. This approach does that and it uses the
variables that the Makefile is already using.

Some background on why kustomize doesn't support env vars in the way
that would be useful to us:
https://github.com/kubernetes-sigs/kustomize/issues/388