Add configuration for public GUE endpoint addresses
This is kind of a wild-assed guess, but here goes: the PureLB clients
start the conversation with the EGW because they're behind firewalls.
That means that the address that they use for the EGW end of the GUE
tunnel needs to be publicly visible. We now have two different public
IP addresses: the Envoy address which is per-loadbalancer, and the GUE
tunnel address which is, I think, per EGW node.

All of these addresses need to be configured somehow so they're
reachable from the public internet. For the Envoy addresses PureLB
should do the job, but it might be overkill for the GUE addresses
because those are going to be configured staticly per node.

The GUE address needs to be sent to PureLB so we'll copy it from the
NodeConfig spec to the LoadBalancer status field.