Move PFC egress setup from the multus bridge to the nodeconfig NIC
This might be temporary, but our current thinking is that having the
egress filter on the physical NIC will be a more rigorous test than
having it on the multus bridge because it will be exposed to a wider
range of packets. If it works here then we might move it back.