Add a ProxyIfindex field to the LoadBalancer status
We need this because we have to figure out what the proxy pod's
ifindex is outside of k8s because of namespace issues, but we need to
use that ifindex inside k8s. We need a way to store the ifindex and
look it up later.

At the moment the field is populated by a Python daemon running on the
host, and it's used by the load balancer controller.