Configure PFC forwarding for endpoints
This commit also moves a bunch of network and PFC config from the
python daemon to the loadbalancer controller. In some ways the python
daemon is more appropriate since it runs on each node in an EGW
cluster, but we plan to experiment with an Envoy pod sidecar config
program so we want the code to be in Go.