Make LoadBalancer.Spec.Endpoints a required field with a default
I want to use Patch() to add endpoints but it looks like there's a
chicken-and-egg problem: you can't "add" an object to an array that
doesn't exist, and the endpoint array doesn't get initialized if it's
omitempty.

I couldn't figure out the kubebuilder annotation to initialize the
endpoints array properly so I'm doing it in the webhook defaulter.