Ensure that an LB's ifindex is set before configuring endpoints
It takes a while for the Envoy pod to launch and for the Python setup
daemon to write the proxy's ifindex to the LB object. If endpoints are
created in the mean time they need to wait until the ifindex is
set. We weren't doing this so we were ending up with endpoints that
didn't have enough info to be able to configure their tunnels.

Now we check and if the ifindex isn't set we back off for a few
seconds and retry.