Don't add the OwningLoadBalancerLabel to endpoints
The web service does that now. We're having a problem with duplicate
endpoints and this is an attempt to solve it. The idea here is that
because we were setting the OwningLoadBalancerLabel in the endpoint
webhook there might be a race condition where we could query for an
LBs endpoints and miss one, which could allow us to create a
duplicate.

I made a change to the web service to set the label when creating the
endpoint resource so now there's no point at which an endpoint is
missing the label.

Since the web service is adding the label the webhook doesn't need to.