Store Account custom resources in their namespaces
We used to store the Account custom resource in the "egw" namespace
but it seems like a better use of namespaces to put the account object
in the same NS as all of the other resources owned by the account.

In this commit we're also switching from Spec fields to Labels to link
one object to another. This makes it easy to query "owner" and
"owned-by". We started this process with remoteendpoints and it works
well.