Add storage in CRDs for Envoy config templates
We're going to try using a template-driven approach to configuring
Envoy. We'll use golang's template package and feed its output into
Marin3r for distribution to Envoy. Still unsolved is how we generate
the templates.

The templates are stored in ServiceGroups so when a user creates a
LoadBalancer it will use the template from its owning ServiceGroup.

The likely flow will be that a user, when creating a ServiceGroup, can
either choose an Envoy template from a library or upload their
own. When the user creates a LoadBalancer that belongs to that group,
its template will be copied into the LoadBalancer CR and processed and
its output will be loaded into a Marin3r EnvoyConfig which Marin3r
will pick up and serve to Envoy.

Changes to a LoadBalancer's template will cause the template to be
re-processed and sent to Envoy. This allows for quick adjustments to
production load balancers.