Take endpoints into account in the LoadBalancer controller
A typical workflow has the LoadBalancer controller firing before any
endpoints exist, but there are other scenarios where an LB might have
endpoints, for example, when clients tweak their LoadBalancer Envoy
template.

Now when we get a LoadBalancer event we Get() the set of endpoints
that belong to that LB. The LB controller doesn't do the PFC setup for
each endpoint but it does ensure that all of the endpoints are
accounted for in the Envoy config. Each endpoint will cause an event
to fire in the endpoint controller, too, so that will set up the PFC.