At startup, initialize data left over from before
Some data, like ifindex values, might change during a system reboot,
so when we're first starting up we can't trust any data that we put
into CR Statuses. We need to wipe out the LoadBalancer Statuses, but
that means that we also need to trigger the Python setup-network
daemon to re-scan the veth and ifindex for each Envoy proxy pod.

This is more ad-hoc that I'd like because the k8s clients that the
controller-runtime manager uses aren't initialized until the manager
starts the controllers, but we want to do this work before any
controller starts.