Make LB deletion more robust
AD found a case where the controller-manager got stuck because an LB
that was being deleted didn't have its corresponding EnvoyConfig. This
caused the controller to return an error before it had removed its
finalizer from the LB so it looped.

Now we remove the finalizer first, and if we can't then it's an
error. Everything else we do in terms of cleanup is "best-effort",
i.e., we try but keep going in case of failure. This should prevent us
from looping if something isn't quite right.