Pass the whole Envoy config through, instead of just parts of it
We weren't passing the Routes, Runtimes, and Secrets. We know that the
Secrets are needed (to configure TLS termination) but we might as well
pass it all through anyway.

Note that Marin3r treats secrets differently than other
EnvoyResources. The other resources are just Envoy config fragments
but Secrets are a Marin3r data type.

For more info:
https://github.com/3scale/marin3r/tree/df8ac5240d37e25dd5d5a9005564d100ae0de98c#secrets