Disable leader elections
We're having a problem where the controller manager crashes and leaves
a lock on the configmap that it uses to elect a leader. This causes it
to go into a crashloop from which it never recovers until someone
manually deletes the configmap.

We'll figure out what exactly is going wrong, but in the mean time
this commit disables the leader election which is OK since we're
running a single node in our EPIC systems.