Use the same TunnelKey for all of a service's tunnels
We found a bug caused by the PFC components using gid/sid to store and
retrieve tunnels keys, which means that multiple tunnels belonging to
the same service would overwrite each other's data.

Now we assign one tunnel key per service and use that for all of that
service's tunnels.