Fix ServiceID allocation
This broke when we split groupID and serviceID. We now store the
ServiceID counter in the ServiceGroup, not the account. Since each
group has its own groupID I think that the group is a better namespace
for the serviceID than the account as it will allow us to have more
services.

There was an unused ProxySnapshotVersions field in the ServiceGroup
status that was used by the erstwhile xds-control-plane but it's no
longer necessary.