Switch sample servicegroup to EDS
We're implementing an Endpoint Discover Service control plane using
our xds-operator so we no longer want to include endpoints in the
clusters section of the template. Now the clusters section points to the
EDS service and there's a new "endpoints" section which the EDS
service will use to format endpoints.