Configure multus with the static IPAM plugin
We were configuring Multus with a host-local IPAM but even when
configured with no range it still checks the pod's IP address against
its subnet. We want to be able to assign any address to any pod. Since
the IPAM isn't managing addresses we can use the static IPAM plugin
which gives us more configuration flexibility. We configure the IPAM
with no addresses to assign because we specify an address for each pod
when we launch it.