Don't fire unneeded webhooks
We've had some problems with clients trying to call webhooks and
getting timeouts. In some cases the webhook wasn't doing anything but
it was active (likely because it was scaffolded that way and I didn't
change it). I believe that we now activate only those webhooks that
are actually doing something.

This should help, especially with remote endpoints since their
activity tends to happen in bursts, for example, when we switch from
"NodePort" to "LoadBalancer" on the client.