Remove artificial constraints on CPU and memory
Since EPIC runs on a dedicated host and the controller-manager is the
core of how it gets configured we want to let the controller-manager
do what it wants to do when it wants to do it.

The controller manager operates in infrequent bursts of activity so we
want to let those bursts finish as quickly as possible.