Add a function to determine an LB's name
The LB's name is a combo of the ServicePrefix name and the "raw" load
balancer name. We need to smash them together because one customer
might have two or more LBs with the same name (but belonging to
different service prefixes) and a customer's LBs all live in the same
k8s namespace so we need to make the service group name into an ersatz
namespace.