Add a DisplayName field to LB
Until now we've been in development mode so we could just use kubectl
to search for the right LB but we're working on epicctl and want to
let users manage LBs with that. We munge the user-provided name to
generate the CR name and parsing is risky so let's cache the
user-provided name so we can just grab it and show it.