Change ProxyInterfaceInfo to allow for multiple envoy pods
We were keying the map with the node address, but now we can have more
than one pod running on a given node, so we're going to use the pod
name as a key and put the node address in the ProxyInterfaceInfo
struct.

At this point we're making the bare minimum changes to work with this
data structure change so this change can't yet handle multiple
pods. Subsequent changes will handle the setup/cleanup for multiple
pods.