Add functions to add and remove finalizers
I'm amazed at how there doesn't seem to be a clear-cut best practice
for something that's so critical to the operation of
controllers. Here's my best guess at how to add and remove
finalizers. I'd rather use Patch() than Update() but can't seem to
find the right incantation.

I copied this code from the eds-server and hope to be able to export
it so we don't need to have it in multiple places.