WIP: rework rep controller to manage the LB tunnel map
Now that we're moving toward multi-node EPIC we've got several
different events that can impact the map of tunnels from EPIC nodes to
client cluster nodes. Two obvious types of events are endpoint changes
and Envoy proxy pod changes. When a new endpoint arrives, for example,
tunnels need to be set up between its node and each proxy's node. On
the other hand, if a new proxy arrives then tunnels need to be set up
from its node to each endpoint node.

Our plan is to have the LB CR be the source of truth, and have the LB
agent controller manage the tunnel setup/teardown. The remote endpoint
controller and pod controller will respond to changes, allocate tunnel
ID's when necessary, and write their changes to the LB status which
will cause the LB agents to fire and implement the changes, such as
setting up new tunnels.

This commit does away with the remoteendpoint_agent_controller because
its functionality will be moved to either the
remoteendpoint_controller or the loadbalancer_agent_controller. This
is a WIP because we implemented what the remoteendpoint_controller
needs to do, but not the LB controller.