Set up privileges for Envoy pods
We enabled RBAC and the Envoy pods failed because they need to read
Calico configuration data at the cluster level in order to set up
their routes.

This commit adds a ClusterRoleBinding to the customer namespace that
binds the namespace's default ServiceAccount to the cluster-level
"manager-role". It also adds kubebuilder code annotations to grant the
new permissions to the manager-role.