Switch from sid/gid to tunnel ID
In the case where Envoy proxy pods on different EPIC nodes need to
connect to a single PureLB node, using groupid/serviceid like we do in
the TABLE-ENCAP isn't good enough because it can't get packets back to
the originating host. Tunnel ID can.

For now we're doing a kludge to split the tunnel id (which is 32 bits)
into two 16 bit fields that we put into the group id and service id
fields. This lets us use the same PFC but will probably be confusing
going forward.