Provide internal service ip CIDR to the Envoy pods
Envoy pods have two interfaces: one for internal traffic and one for
TrueIngress traffic. We set up the routing so TrueIngress is the
default but we know that k8s' internal service traffic won't use
TrueIngress so now we add a route to send that traffic over the
internal network interface.

The service network CIDR must be configured in the epic singleton
custom resource. The snap packaging does this.