Incorporate external-dns' Endpoint CRD/struct into LB and LBSG
With some changes to external-dns[1] it's possible to use your own
custom resources as external-dns "sources". This commit incorporates
parts of external-dns's CRD into ours so we can do that with our
LoadBalancer CRD. I did it this way and not by "go get" because
external-dns has a very long list of dependencies and "go get" failed
with some sort of DLL-hell conflict when I tried to do that.  This
approach should have a low impact on resource-model since it's just
adding a couple of structs and not dozens of module dependencies.

I copied external-dns' endpoint/endpoint.go and its tests into this
project, and added external-dns' endpoint/labels.go as dns_labels.go
since we already had a file named labels.go. I cut the top-level
structs (DNSEndpoint and friends) out of endpoint.go and changed the
package to match ours. With these changes LoadBalancer can now be an
external-dns source.

[1] https://github.com/acnodal/external-dns/tree/dynamic-crd