Clean per-node Envoy pod info when restarting node agents
When a node agent starts up it can't trust the Envoy pod info for the
pods running on its host. That info might have changed while the agent
wasn't running. We need to zero the info and "nudge" the pods so the
python daemon will update the pod annotations with the current info.

This is basically the same thing that we do when the allocator starts,
only the scope is limited to just the pods running on this node.