Fix a bug that dragged prometheus in as a dependency
We don't need it, but emacs' LSP can be a little aggressive about
trying to find dependencies. In this case we had used a
variable ("log") before initializing it so LSP "helpfully" found a
package called github.com/prometheus/common/log which "fixed" the
problem, except it wasn't really a fix.