Switch from k8s Deployment to Marin3r EnvoyDeployment
This was a lot of work, and I hope it pays off in easier long-term
maintenance. We no longer build and apply a k8s Deployment with a
bunch of reverse-engineered hacks to be able to connect to
Marin3r. Now we apply an EnvoyDeployment and Marin3r makes sure that
the Envoy pod has the bootstrap that it needs to be able to query
Marin3r's XDS for its config.

At least this reduces the coupling between EPIC and Marin3r by
reducing the number of undocumented Marin3r features that we use.