Add the host IP address to envoy pod EnvVars
We'll use it to set up a route inside the pod. We need an explicit
route for the host IP to handle the case where kubelet is running on
the same host as an Envoy container and wants to probe the container
for liveness and/or readiness. In that case the packets come from the
host IP so we need them to return via the pod's eth0, not the pod's
PFC interface.