Move Multus route addition to the pod agent controller
This is because we want to attract traffic to the Envoy pod only once
the PFC tunnel and upstream are both known to be ready. We previously
set up a kubelet readiness probe for this purpose but we're not using
the results since readiness only affects whether a pod is considered a
valid endpoint for a Service.

We also have a liveness probe inherited from Marin3r and that will
restart the Envoy pod when it fails. That just asks Envoy if it's
live, though, so it doesn't say anything about the PFC or the upstream
cluster.