First attempt at the server side of our Gateway implementation
This is an obnoxiously large commit but I was in "furious hack" mode
for much of this development. When I got out from under the deadlines
I had hopelessly messed up the git history so I just said "mea culpa"
and squashed the commits.

This is mostly in parallel with our previous LoadBalancer
implementation. The pod controllers are shared because both
implementations use Marin3r to manage Envoy, but other than that
they're disjoint.