Share a common FinalizerName among controllers
There doesn't seem to be a reason for them all to have their own
name. In cases like the agent controllers where they run on each node
then the names need to include the node names but for traditional
controllers that run a single instance we can use the same finalizer
name for all.