Provide the pod cidr to envoy deployments
We've been using calicoctl to determine the pod cidr from inside the
Envoy pods, but now we're switching to Flannel. In any case, we should
do this in a CNI-independent way.

Now we get the pod cidr from the node spec and put that into the envoy
deploy env along with the service cidr.