Remove installation templates
This is a bummer. I'd like to have the templates here, and then have
devops download them to the host and pass them through the template
processor, but for some reason that was far more difficult than I
thought it should be. Ansible seems to not like to run things on the
host, but I finally got it working with the Ansible in Debian
11. Unfortunately it didn't work with the older Ansible in Ubuntu
20.04 which Adam uses. So I had to fall back to storing these
templates in the Ansible playbook. Grrrr.