Move ServicePrefix public pool info into AddressPool object
This ended up being a much bigger change than I anticipated. It
started with me wanting to add the "alt" address to the ipset groups
and the multus routes. Unfortunately, the code to do that belonged to
the SP Spec so I moved it into the new AddressPool so I could use the
same code for both pools. This was a much more intrusive change than I
expected.