Provide a default DNS record type when creating an Endpoint
I'm trying to make template maintenance easier and it looks like the
only difference between IPV4 and IPV6 templates is the DNS record
type ("AAAA" vs "A"). Now if the LBSG doesn't specify a record type
the code fills in either A or AAAA depending on the public IP's
family.

This will hopefully enable us to use a single template for IPV4 and
IPV6 which will reduce the complexity of rolling out template changes.