Sort HTTPRouteMatch objects by complexity
Simpler criteria are easier to match so they need to be checked later
than more complex criteria. You might have two criteria with the same
path but one of them also has a header. The one with two different
matches needs to be first because otherwise it will never even get
checked because the simpler one will match in every case that the
complex one would.

This adds another rule to sortRouteRules. I doubt that it will be the
last but at least we've got a place to add these rules.