Split multi-match rules into multiple single-match rules
We do this because we need to re-order the matches from most-specific
to least-specific for Envoy to do its job. A single rule might have
several matches, but some of the matches might be more specific and
some might be less specific. In this case some of the matches need to
be early in the list, and some need to be later, but if they all
belong to one rule then they end up in the same place.

We're able to do this because matches within a rule are "OR-ed", i.e.,
the rule matches if *any* of its matches matches. We can therefore
take a rule with two matches and turn it into two rules, each with one
match. This allows us to sort them into different places in the list.