Don't pass POD_CIDR into Envoy deployments
It turns out that the pod CIDR is per-node so this didn't work once
we started running on multi-node clusters. Actually, what's worse is
that it sorta worked while the cluster was mostly empty since
everything ended up on the same node, but then it started to fail as
the cluster got more full and pods ended up on different nodes.