Remove obsolete LoadBalancer code
We switched to GWProxy and friends a while ago. We're still using
RemoteEndpoint but no longer use LoadBalancer for anything. This
commit removes all uses of LoadBalancer but doesn't remove the CRD
itself. We'll make sure that the other components (web service,
control plane, etc) are no longer using it first.