Use "raw" objects to build the GWP tunnel map patch
The problem with using CRs and then Marshalling them is that you get a
bunch of default values that you probably don't want so your patch
ends up changing things that you weren't expecting.

My current thinking is that if you're using Patch() then you probably
want to make only a small change so it's best to build the patch with
map[string]interface{}{} instead of custom resources since that way
you can control exactly what's in the patch.