Upgrade gateway-api to a patched version of v0.5.0-rc1
This version has URL rewriting which we want to implement.

Unfortunately, the URLRewrite feature is in the experimental channel
so it doesn't work with code like ours that includes the gateway API
by reference. They've evidently got some kludges in a shell script
somewhere but we don't have that option.

So the version that we're "replace"-ing promotes the URLRewrite to be
part of the normal channel or whatever they call it.