Combine Route Rules
We had a problem with multiple Routes attaching to a single
Listener. Say you've got two upstream clusters. Each will have a Route
and they'll share a Listener. Before this change, we'd end up with the
Envoy config having two routes with the same matches so the first one
would always be chosen and the second one would never be chosen.

We want one route with multiple weighted_clusters so Envoy can
round-robin around all of them.

Signed-off-by: Toby Cabot <toby@acnodal.io>