Handle non-HTTP GWRoutes
When we first implemented the Gateway spec we supported only
HTTPRoutes but we had a place to put other types. Unfortunately, as
usually happens, a bunch of assumptions crept into the code, usually
things like route.Spec.HTTP.Rules which will crash if route.Spec.HTTP
is nil (i.e., if it's not an HTTPRoute).

This patch does the bare minimum to fix the code paths so that
non-HTTP routes don't crash. They don't actually work, but this is
progress.