Pre-process only HTTP routes, not other types
HTTP routes (i.e., GWRoutes containing an HTTPRoute) need to be
pre-processed to map the Gateway model onto Envoy's model. We do this
in Go because it's easier for me to read and understand Go code than
Go's template language.

Now that we're adding TCP routes it turns out that they don't need to
be preprocessed so now we build two lists (one of HTTP routes and one
of non-HTTP routes) and preprocess only the HTTP list.