Tweak logging configuration and root command structure
It turns out that our logging config code was in one of our commands
so we weren't always initializing logging properly. Now we wait until
Cobra has been initialized so we can read the debug flag.

I also added the log as an explicit parameter to a few functions so we
reduced our use of ctrl.Log(), and our coupling to the ctrl package.

Signed-off-by: Toby Cabot <toby@acnodal.io>