Fix some code that assumed that it was running in the cluster
We want to be able to run on developers' machines as well as inside
the cluster, so we can use tools like debuggers. Unfortunately there
are still some bugs that prevent us running at the command line but
this fix gets us part of the way there.

The "run-manager" Makefile target can be used to run the manager at
the command line.